pyproject.toml
[tool.poetry]
name = "baddns"
version = "1.12.0"
description = "Check subdomains for subdomain takeovers and other DNS tomfoolery"
authors = ["liquidsec <paul.mueller08@gmail.com>"]
repository = "https://github.com/blacklanternsecurity/baddns"
homepage = "https://github.com/blacklanternsecurity/baddns"
documentation = "https://www.blacklanternsecurity.com/baddns/"
license = "GPL-3.0"
readme = "README.md"
include = ["signatures/**/*"]

[tool.poetry.dependencies]
python = "^3.9"
pyyaml = "^6.0.1"
dnspython = "^2.4.2"
colorama = "^0.4.6"
python-whois = "^0.9.6"
tldextract = "^5.1.1"
python-dateutil = "2.9.0.post0"
httpx = ">=0.27.0,<0.29.0"

[tool.poetry.group.dev.dependencies]
requests-mock = "^1.11.0"
pytest = "^7.4.0"
pytest-cov = "^4.0.0"
pytest-asyncio = "^0.21.1"


[tool.poetry.group.docs.dependencies]
mkdocs = "^1.5.3"
mkdocs-material = "^9.5.8"

[tool.black]
line-length = 119

[tool.poetry.scripts]
baddns = 'baddns.cli:main'

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning"
]

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry-dynamic-versioning]
enable = true
metadata = true
format = '1.12.{distance}'
